{"schema":"decodex/commit/1","summary":"Consolidate cohesive account panel view modules","authority":"manual"}
