--- apps/decodex-app/Sources/DecodexApp/AccountPanelLayoutMetrics.swift
@@ -1,57 +1,0 @@
-import SwiftUI
-
-extension AccountPanelView {
-	var accountListContentHeight: CGFloat {
-		store.accounts.reduce(CGFloat(1)) { total, account in
-			total + accountRowHeight(for: account)
-		}
-	}
-
-	var accountListViewportHeight: CGFloat {
-		min(accountListContentHeight, accountListAvailableHeight)
-	}
-
-	var accountListNeedsScrolling: Bool {
-		accountListContentHeight > accountListAvailableHeight + 1
-	}
-
-	var accountListAvailableHeight: CGFloat {
-		let visibleHeight = AccountPanelLayout.activeScreenVisibleHeight()
-		let availableHeight = visibleHeight - accountPanelChromeHeight
-		let minimumHeight = min(
-			AccountPanelLayout.minimumScrollableListHeight,
-			max(140, visibleHeight * 0.42)
-		)
-
-		return max(minimumHeight, availableHeight)
-	}
-
-	var accountPanelChromeHeight: CGFloat {
-		var height = AccountPanelLayout.screenVerticalMargin
-			+ AccountPanelLayout.panelVerticalPadding
-			+ AccountPanelLayout.headerHeight
-			+ AccountPanelLayout.accountSummaryHeight
-			+ AccountPanelLayout.sectionSpacing
-
-		if telemetryMatrixIsVisible {
-			height += AccountPanelLayout.sectionSpacing + telemetryMatrixHeight
-		}
-		if store.notice != nil {
-			height += AccountPanelLayout.sectionSpacing + AccountPanelLayout.noticeHeight
-		}
-		if store.accountList?.usageProbeError != nil {
-			height += AccountPanelLayout.sectionSpacing + AccountPanelLayout.noticeHeight
-		}
-
-		return height
-	}
-
-	var accountScrollProbe: some View {
-		GeometryReader { proxy in
-			Color.clear.preference(
-				key: AccountScrollOffsetPreferenceKey.self,
-				value: proxy.frame(in: .named(AccountPanelLayout.accountListScrollSpace)).minY
-			)
-		}
-	}
-}
