--- apps/decodex-app/Sources/DecodexApp/AccountPanelTelemetryMetrics.swift
@@ -1,33 +1,0 @@
-import SwiftUI
-
-extension AccountPanelView {
-	var accountProfileAggregate: AccountProfileAggregate? {
-		AccountProfileAggregate.make(accounts: store.accounts)
-	}
-
-	var telemetryMatrixIsVisible: Bool {
-		accountProfileAggregate != nil
-			|| store.accountList?.usageEstimate != nil
-	}
-
-	var telemetryMatrixHeight: CGFloat {
-		var rows = [CGFloat]()
-		if accountProfileAggregate != nil {
-			rows.append(AccountPanelLayout.telemetryProfileHeight)
-		}
-		if let estimate = store.accountList?.usageEstimate {
-			rows.append(
-				estimate.accountEstimateCount < estimate.accountCount
-					? AccountPanelLayout.telemetryPoolMeasuredHeight
-					: AccountPanelLayout.telemetryPoolHeight
-			)
-		}
-		guard rows.isEmpty == false else {
-			return 0
-		}
-
-		return AccountPanelLayout.telemetryVerticalPadding
-			+ rows.reduce(0, +)
-			+ CGFloat(rows.count - 1) * AccountPanelLayout.telemetryRowSpacing
-	}
-}
